iOS Retrieves Last Sleep Session and Responds to Watch
Added method to package a dictionary with the most recent sleep session
data. This is returned to the watch app where it is populated into the
milestone table. WCSession object was added but this should be removed
when implementing a WCSession wrapper class.

Some irregularities led me to implement a check for the WCSession's
reachable property. The should continue in the future.

Finally the milestone table view will not populate on its own. Currently
in a debug mode while implementing final steps.

-- populateDictionaryWithSleepSessionData (added)
-- updateMilestoneTableData (added)
-- requestMostRecentSleepSessionFromiOSApp (major rewrite)
-- convertAndPopulateSleepSessionDataForMilestone (added)