Add Support For Displaying Multi-Sleep Session
Closes #4

Removed much of the core data code as the response time was
not preferred to storing the most recent sleep session to the
Apple Watch's documents directory.

* Merged Watch Connectivity code to support a "Response" key
when determining how to execute request.
* Utility now parses data from Apple Watch's directory as opposed
to a previously used core data directory.
* reloadMilestoneInterfaceData is called to reload the interface to
present user with the most recent sleep session without any
interaction.
* Organized Sleep Delegate Functions to simply accept proposed
sleep time if it was accepted, then call performSleepSessionCloseout.