Manually Push Last Sleep Session to iOS App
Closes #28

On rare occasions the iOS app would not receive the sleep session from the WatchOS application. This would cause inconsistencies in the statistics and reporting of the device. Users are now able to access a menu icon from the second page of the iOS app to manually resend the sleep session to the iOS device again.

* Documented new variables
* Moved variable and WCSession initialization to init method
* Removed redefinition of lastSleepSession object
* Set the creationDate to reflect the date sleep occurred and not the date sleep was sent to iOS