Foundation For Heart Rate Graphing
Began development on #40

Originally implemented the FSCharts library to allow for graphing heart rate data in the sleep session detail view. This was later decided against, but not removed, in an effort to create the charts in-house. Created the HeartRateChart class and began initial development of this class to allow for a heart rate chart to be displayed once a sleep session has been saved.

* Added new chart class, HeartRateChart
* Installed FSCharts and pointed to my forked version

Resource Installed:
FSLineChart
https://github.com/ArthurGuibert/FSLineChart