Added Foundation For Supporting Local Notifications
Began development on issue #9

This commit provides a strong foundation for supporting local notifications on the apple watch extension of the Sleepy app. Currently the push notification will alert the user that they have marked themselves as awake, but have not fully ended the sleep session. The notification will present the user with options to end the session, snooze the notification, or dismiss.

* Added new page to the on boarding process to request user to allow notifications.
* Created stringWithUUID to generate a unique UUID for the notification identifier.
* Created 4 constants to retain string values and time interval for end sleep session notification.
* Created 3 constants to retain string values for new on boarding page.