Address Issues with NSTimer Not Triggering When Application Enters Background
Closes #11

When testing commit 76e1822, it was observed that the timer would not trigger nor continue its scheduled time if the application entered the background. This commit changes the way we trigger the methods to adjust the menu icons that are displayed to the user by storing the values in the user defaults and checking against this when the user is in a sleep session.

* Created method determineMenuIconsToDisplay
* Updated outdated reference to Health.plist as we now use SavedSleepSession.plist
* Added method isUserAwake
* Created writeRemoveDeferredSleepOptionDate to persist date which deferred sleep option should be removed
* Removed old references to NSTimer solution from commit 76e1822