Support For Dismissing Proposed Sleep Interface
Closes #49

Currently watchOS requires an interface presented with presentControllerWithName: to include a "cancel" button. If a user canceled this interface previously, it would cause the sleep session to fall out of sync which would lead to a crash. This is now resolved by presenting the user with a menu icon that will simply prompt the user to repeat the proposed sleep interface once again.

* Created new method prepareMenuIconsForUserDismissedProposedSleepInterface to allow user to display the proposed sleep time interface
* sleepDidStopMenuButton called new menu icon method